Task: BE #7 - Add Validation

+++ web_app/services/users/user_service.py
@@ -2,6 +2,7 @@ from fastapi import Depends
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from web_app.db.postgres_helper import postgres_helper as pg_helper
+from web_app.exceptions.permission import PermissionDeniedException
 from web_app.exceptions.users import (
     UserEmailNotFoundException,
     UserIdAlreadyExistsException,
@@ -45,16 +46,22 @@ class UserService:
         return await self.user_repository.create_obj(new_user)
 
     async def update_user(
-            self, user_id: int, user_update: UserUpdateRequestModel
+        self,
+        user_id: int,
+        user_update: UserUpdateRequestModel,
+        current_user: User
     ) -> User:
-        user = await self.get_user_by_id(user_id)
+        if current_user.id != user_id:
+            raise PermissionDeniedException()
         if user_update.first_name:
-            user.first_name = user_update.first_name
+            current_user.first_name = user_update.first_name
         if user_update.last_name:
-            user.last_name = user_update.last_name
-        return await self.user_repository.update_obj(user, user_id)
+            current_user.last_name = user_update.last_name
+        return await self.user_repository.update_obj(current_user, user_id)
 
-    async def delete_user(self, user_id: int):
+    async def delete_user(self, user_id: int, current_user: User) -> None:
+        if current_user.id != user_id:
+            raise PermissionDeniedException()
         user = await self.get_user_by_id(user_id)
         await self.user_repository.delete_obj(user.id)
 
@@ -72,8 +79,7 @@ class UserService:
         new_user = User(
             first_name=email.split('@')[0],
             last_name="",
-            email=email,
-            password=PasswordManager.return_default_password()
+            email=email
         )
         return await self.user_repository.create_obj(new_user)
